+++ src/core/constraint_template.jl
@@ -10,6 +10,82 @@
 # Constraint templates should always be defined over "GenericGasModel"
 # and should never refer to model variables
 
+###############################################################################################
+# Templates for constraints associated with resistors 
+###############################################################################################
+
+" Template: Constraint on mass flow across a pipe"
+function constraint_resistor_mass_flow(gm::GenericGasModel, n::Int, k)
+    pipe           = ref(gm,n,:connection,k)
+    w              = ref(gm,n,:w)[k]
+    f_min          = calc_resistor_fmin(gm, n, k, w)
+    f_max          = calc_resistor_fmax(gm, n, k, w)
+    constraint_resistor_mass_flow(gm, n, k, f_min, f_max)
+end
+constraint_resistor_mass_flow(gm::GenericGasModel, k::Int) = constraint_resistor_mass_flow(gm, gm.cnw, k)
+
+" Template: Constraint on flow across a resistor where the flow is constrained to one direction as defined by data attribute directed"
+function constraint_resistor_mass_flow_directed(gm::GenericGasModel, n::Int, k)
+    pipe           = ref(gm,n,:connection,k)
+    direction      = pipe["directed"]
+    w              = ref(gm,n,:w)[k]
+    f_min          = (direction == 1) ? 0 : calc_resistor_fmin(gm, n, k, w)
+    f_max          = (direction == 1) ? calc_resistor_fmax(gm, n, k, w) : 0
+    constraint_resistor_mass_flow_directed(gm, n, k, f_min, f_max)
+end
+constraint_resistor_mass_flow_directed(gm::GenericGasModel, k::Int) = constraint_resistor_mass_flow_directed(gm, gm.cnw, k)
+
+"Template: Pressure drop across resistor with on/off direction variables"
+function constraint_resistor_pressure(gm::GenericGasModel, n::Int, k)
+    pipe           = ref(gm, n, :connection, k)
+    i              = pipe["f_junction"]
+    j              = pipe["t_junction"]
+    pd_max         = ref(gm,n,:pd_max)[k]
+    pd_min         = ref(gm,n,:pd_min)[k]
+    constraint_resistor_pressure(gm, n, k, i, j, pd_min, pd_max)
+end
+constraint_resistor_pressure(gm::GenericGasModel, k::Int) = constraint_resistor_pressure(gm, gm.cnw, k)
+
+" Template: Constraint on pressure drop across a resistor where the flow is constrained to one direction as defined by data attribute directed"
+function constraint_resistor_pressure_directed(gm::GenericGasModel, n::Int, k)
+    pipe           = ref(gm, n, :connection, k)
+    i              = pipe["f_junction"]
+    j              = pipe["t_junction"]
+    direction      = pipe["directed"]
+    pd_max         = (direction == 1) ? ref(gm,n,:pd_max)[k] : min(0, ref(gm,n,:pd_max)[k])
+    pd_min         = (direction == 1) ? max(0, ref(gm,n,:pd_min)[k]) : ref(gm,n,:pd_min)[k]
+    constraint_resistor_pressure_directed(gm, n, k, i, j, pd_min, pd_max)
+end
+constraint_resistor_pressure_directed(gm::GenericGasModel, k::Int) = constraint_resistor_pressure_directed(gm, gm.cnw, k)
+
+"Template: Weymouth equation for defining the relationship between pressure drop and flow across a resistor "
+function constraint_resistor_weymouth(gm::GenericGasModel, n::Int, k)
+    pipe    = ref(gm,n,:connection,k)
+    i       = pipe["f_junction"]
+    j       = pipe["t_junction"]
+    mf      = gm.ref[:nw][n][:max_mass_flow]
+    w       = ref(gm,n,:w)[k]
+    pd_max  = ref(gm,n,:pd_max)[k]
+    pd_min  = ref(gm,n,:pd_min)[k]
+    f_min   = calc_resistor_fmin(gm, n, k, w)
+    f_max   = calc_resistor_fmax(gm, n, k, w)
+    constraint_resistor_weymouth(gm, n, k, i, j, f_min, f_max, w, pd_min, pd_max)
+end
+constraint_resistor_weymouth(gm::GenericGasModel, k::Int) = constraint_resistor_weymouth(gm, gm.cnw, k)
+
+"Template: Weymouth equation for defining the relationship between pressure drop and flow across a resistor where flow is constrained in one direction"
+function constraint_resistor_weymouth_directed(gm::GenericGasModel, n::Int, k)
+    pipe      = ref(gm,n,:connection,k)
+    i         = pipe["f_junction"]
+    j         = pipe["t_junction"]
+    w         = ref(gm, n, :w)[k]
+    direction = pipe["directed"]
+    f_min     = calc_resistor_fmin(gm, n, k, w)
+    f_max     = calc_resistor_fmax(gm, n, k, w)
+    constraint_resistor_weymouth_directed(gm, n, k, i, j, w, f_min, f_max, direction)
+end
+constraint_resistor_weymouth_directed(gm::GenericGasModel, k::Int) = constraint_resistor_weymouth_directed(gm, gm.cnw, k)
+
 #################################################################################################
 # Templates for constraints associated with pipes
 #################################################################################################
@@ -24,6 +100,7 @@ function constraint_pipe_mass_flow(gm::GenericGasModel, n::Int, k)
 end
 constraint_pipe_mass_flow(gm::GenericGasModel, k::Int) = constraint_pipe_mass_flow(gm, gm.cnw, k)
 
+
 " Template: Constraint on flow across a pipe where the flow is constrained to one direction as defined by data attribute directed"
 function constraint_pipe_mass_flow_directed(gm::GenericGasModel, n::Int, k)
     pipe           = ref(gm,n,:connection,k)
@@ -34,6 +111,8 @@ function constraint_pipe_mass_flow_directed(gm::GenericGasModel, n::Int, k)
     constraint_pipe_mass_flow_directed(gm, n, k, f_min, f_max)
 end
 constraint_pipe_mass_flow_directed(gm::GenericGasModel, k::Int) = constraint_pipe_mass_flow_directed(gm, gm.cnw, k)
+
+
 
 "Template: Pressure drop across pipes with on/off direction variables"
 function constraint_pipe_pressure(gm::GenericGasModel, n::Int, k)
@@ -46,6 +125,8 @@ function constraint_pipe_pressure(gm::GenericGasModel, n::Int, k)
 end
 constraint_pipe_pressure(gm::GenericGasModel, k::Int) = constraint_pipe_pressure(gm, gm.cnw, k)
 
+
+
 " Template: Constraint on pressure drop across a pipe where the flow is constrained to one direction as defined by data attribute directed"
 function constraint_pipe_pressure_directed(gm::GenericGasModel, n::Int, k)
     pipe           = ref(gm, n, :connection, k)
@@ -57,6 +138,8 @@ function constraint_pipe_pressure_directed(gm::GenericGasModel, n::Int, k)
     constraint_pipe_pressure_directed(gm, n, k, i, j, pd_min, pd_max)
 end
 constraint_pipe_pressure_directed(gm::GenericGasModel, k::Int) = constraint_pipe_pressure_directed(gm, gm.cnw, k)
+
+
 
 "Template: Constraints on flow across an expansion pipe with on/off direction variables "
 function constraint_pipe_mass_flow_ne(gm::GenericGasModel, n::Int, k)
@@ -83,16 +166,16 @@ end
 constraint_pipe_pressure_ne(gm::GenericGasModel, k::Int) = constraint_pipe_pressure_ne(gm, gm.cnw, k)
 
 " Template: Constraint on pressure drop across an expansion pipe where the flow is constrained to one direction as defined by data attribute directed"
-function constraint_pressure_ne_directed(gm::GenericGasModel, n::Int, k)
+function constraint_pipe_pressure_ne_directed(gm::GenericGasModel, n::Int, k)
     pipe           = ref(gm, n, :ne_connection, k)
     i              = pipe["f_junction"]
     j              = pipe["t_junction"]
     direction      = pipe["directed"]
     pd_max         = ref(gm,n,:pd_max_ne)[k]
     pd_min         = ref(gm,n,:pd_min_ne)[k]
-    constraint_pressure_ne_directed(gm, n, k, i, j, pd_min, pd_max, direction)
+    constraint_pipe_pressure_ne_directed(gm, n, k, i, j, pd_min, pd_max, direction)
 end
-constraint_pressure_ne_directed(gm::GenericGasModel, k::Int) = constraint_pressure_ne_directed(gm, gm.cnw, k)
+constraint_pipe_pressure_ne_directed(gm::GenericGasModel, k::Int) = constraint_pipe_pressure_ne_directed(gm, gm.cnw, k)
 
 "Template: Constraints on flow across an expansion pipe where the flow is constrained to one direction "
 function constraint_pipe_mass_flow_ne_directed(gm::GenericGasModel, n::Int, k)
@@ -108,7 +191,7 @@ end
 constraint_pipe_mass_flow_ne_directed(gm::GenericGasModel, k::Int) = constraint_pipe_mass_flow_ne_directed(gm, gm.cnw, k)
 
 "Template: Weymouth equation for defining the relationship between pressure drop and flow across a pipe "
-function constraint_weymouth(gm::GenericGasModel, n::Int, k)
+function constraint_pipe_weymouth(gm::GenericGasModel, n::Int, k)
     pipe    = ref(gm,n,:connection,k)
     i       = pipe["f_junction"]
     j       = pipe["t_junction"]
@@ -118,20 +201,26 @@ function constraint_weymouth(gm::GenericGasModel, n::Int, k)
     pd_min  = ref(gm,n,:pd_min)[k]
     f_min   = calc_pipe_fmin(gm, n, k, w)
     f_max   = calc_pipe_fmax(gm, n, k, w)
-    constraint_weymouth(gm, n, k, i, j, f_min, f_max, w, pd_min, pd_max)
+    constraint_pipe_weymouth(gm, n, k, i, j, f_min, f_max, w, pd_min, pd_max)
 end
-constraint_weymouth(gm::GenericGasModel, k::Int) = constraint_weymouth(gm, gm.cnw, k)
+constraint_pipe_weymouth(gm::GenericGasModel, k::Int) = constraint_pipe_weymouth(gm, gm.cnw, k)
+
+
 
 "Template: Weymouth equation for defining the relationship between pressure drop and flow across a pipe where flow is constrained in one direction"
-function constraint_weymouth_directed(gm::GenericGasModel, n::Int, k)
+function constraint_pipe_weymouth_directed(gm::GenericGasModel, n::Int, k)
     pipe      = ref(gm,n,:connection,k)
     i         = pipe["f_junction"]
     j         = pipe["t_junction"]
     w         = ref(gm, n, :w)[k]
     direction = pipe["directed"]
-    constraint_weymouth_directed(gm, n, k, i, j, w, direction)
+    f_min     = calc_pipe_fmin(gm, n, k, w)
+    f_max     = calc_pipe_fmax(gm, n, k, w)
+    constraint_pipe_weymouth_directed(gm, n, k, i, j, w, f_min, f_max, direction)
 end
-constraint_weymouth_directed(gm::GenericGasModel, k::Int) = constraint_weymouth_directed(gm, gm.cnw, k)
+constraint_pipe_weymouth_directed(gm::GenericGasModel, k::Int) = constraint_pipe_weymouth_directed(gm, gm.cnw, k)
+
+
 
 "Template: Constraint associatd with turning off flow depending on the status of expansion pipes"
 function constraint_pipe_ne(gm::GenericGasModel, n::Int, k)
@@ -144,7 +233,7 @@ end
 constraint_pipe_ne(gm::GenericGasModel, k::Int) = constraint_pipe_ne(gm, gm.cnw, k)
 
 " Template: Weymouth equation for expansion pipes "
-function constraint_weymouth_ne(gm::GenericGasModel,  n::Int, k)
+function constraint_pipe_weymouth_ne(gm::GenericGasModel,  n::Int, k)
     pipe           = gm.ref[:nw][n][:ne_connection][k]
     i              = pipe["f_junction"]
     j              = pipe["t_junction"]
@@ -154,12 +243,12 @@ function constraint_weymouth_ne(gm::GenericGasModel,  n::Int, k)
     pd_min         = ref(gm,n,:pd_min_ne)[k]
     f_min          = calc_pipe_ne_fmin(gm, n, k, w)
     f_max          = calc_pipe_ne_fmax(gm, n, k, w)
-    constraint_weymouth_ne(gm, n, k, i, j, w, f_min, f_max, pd_min, pd_max)
+    constraint_pipe_weymouth_ne(gm, n, k, i, j, w, f_min, f_max, pd_min, pd_max)
 end
-constraint_weymouth_ne(gm::GenericGasModel, k::Int) = constraint_weymouth_ne(gm, gm.cnw, k)
+constraint_pipe_weymouth_ne(gm::GenericGasModel, k::Int) = constraint_pipe_weymouth_ne(gm, gm.cnw, k)
 
 " Template: Weymouth equation for expansion pipes where flow is restricted to one direction "
-function constraint_weymouth_ne_directed(gm::GenericGasModel,  n::Int, k)
+function constraint_pipe_weymouth_ne_directed(gm::GenericGasModel,  n::Int, k)
     pipe           = gm.ref[:nw][n][:ne_connection][k]
     i              = pipe["f_junction"]
     j              = pipe["t_junction"]
@@ -168,9 +257,11 @@ function constraint_weymouth_ne_directed(gm::GenericGasModel,  n::Int, k)
     direction      = pipe["directed"]
     pd_max         = ref(gm,n,:pd_max_ne)[k]
     pd_min         = ref(gm,n,:pd_min_ne)[k]
-    constraint_weymouth_ne_directed(gm, n, k, i, j, w, pd_min, pd_max, direction)
+    f_min          = calc_pipe_ne_fmin(gm, n, k, w)
+    f_max          = calc_pipe_ne_fmax(gm, n, k, w)
+    constraint_pipe_weymouth_ne_directed(gm, n, k, i, j, w, pd_min, pd_max, f_min, f_max, direction)
 end
-constraint_weymouth_ne_directed(gm::GenericGasModel, k::Int) = constraint_weymouth_ne_directed(gm, gm.cnw, k)
+constraint_pipe_weymouth_ne_directed(gm::GenericGasModel, k::Int) = constraint_pipe_weymouth_ne_directed(gm, gm.cnw, k)
 
 #################################################################################################
 # Templates for constraints associated with junctions