+++ src/form/lp.jl
@@ -28,17 +28,31 @@ end
 ######################################################################################################
 
 "Constraint: Weymouth equation--not applicable for LP models"
-function constraint_weymouth(gm::GenericGasModel{T}, n::Int, k, i, j, f_min, f_max, w, pd_min, pd_max) where T <: AbstractLPForm
+function constraint_pipe_weymouth(gm::GenericGasModel{T}, n::Int, k, i, j, f_min, f_max, w, pd_min, pd_max) where T <: AbstractLPForm
+    #TODO we could think about putting a polyhendra around the weymouth
+end
+
+"Constraint: Weymouth equation--not applicable for LP models"
+function constraint_resistor_weymouth(gm::GenericGasModel{T}, n::Int, k, i, j, f_min, f_max, w, pd_min, pd_max) where T <: AbstractLPForm
     #TODO we could think about putting a polyhendra around the weymouth
 end
 
 "Constraint: Weymouth equation with one way direction--not applicable for LP models"
-function constraint_weymouth_directed(gm::GenericGasModel{T}, n::Int, k, i, j, w, direction) where T <: AbstractLPForm
+function constraint_pipe_weymouth_directed(gm::GenericGasModel{T}, n::Int, k, i, j, w, f_min, f_max, direction) where T <: AbstractLPForm
+    #TODO we could think about putting a polyhendra around the weymouth
+end
+
+"Constraint: Weymouth equation with one way direction--not applicable for LP models"
+function constraint_resistor_weymouth_directed(gm::GenericGasModel{T}, n::Int, k, i, j, w, f_min, f_max, direction) where T <: AbstractLPForm
     #TODO we could think about putting a polyhendra around the weymouth
 end
 
 " Constraint: constraints on pressure drop across where direction is constrained"
 function constraint_pipe_pressure_directed(gm::GenericGasModel{T}, n::Int, k, i, j, pd_min, pd_max) where T <: AbstractLPForm
+end
+
+" Constraint: constraints on pressure drop across where direction is constrained"
+function constraint_resistor_pressure_directed(gm::GenericGasModel{T}, n::Int, k, i, j, pd_min, pd_max) where T <: AbstractLPForm
 end
 
 " Constraint: Constraint on pressure drop across a short pipe--not applicable for LP models"
@@ -66,15 +80,19 @@ function constraint_on_off_control_valve_pressure_directed(gm::GenericGasModel{T
 end
 
 "Constraint: Weymouth equation--not applicable for MIP models--not applicable for LP models"
-function constraint_weymouth_ne(gm::GenericGasModel{T},  n::Int, k, i, j, w, f_min, f_max, pd_min, pd_max) where T <: AbstractLPForm
+function constraint_pipe_weymouth_ne(gm::GenericGasModel{T},  n::Int, k, i, j, w, f_min, f_max, pd_min, pd_max) where T <: AbstractLPForm
 end
 
 " Constraint: Pressure drop across an expansion pipe when direction is constrained--not applicable for LP models"
-function constraint_pressure_drop_ne_directed(gm::GenericGasModel{T}, n::Int, k, i, j, yp, yn) where T <: AbstractLPForm
+function constraint_pipe_pressure_drop_ne_directed(gm::GenericGasModel{T}, n::Int, k, i, j, yp, yn) where T <: AbstractLPForm
+end
+
+" Constraint: Pressure drop across an expansion pipe when direction is constrained--not applicable for LP models"
+function constraint_resistor_pressure_drop_ne_directed(gm::GenericGasModel{T}, n::Int, k, i, j, yp, yn) where T <: AbstractLPForm
 end
 
 "Constraint: Weymouth equation--not applicable for MIP models--not applicable for LP models"
-function constraint_weymouth_ne_directed(gm::GenericGasModel{T},  n::Int, k, i, j, w, pd_min, pd_max, direction) where T <: AbstractLPForm
+function constraint_pipe_weymouth_ne_directed(gm::GenericGasModel{T},  n::Int, k, i, j, w, pd_min, pd_max, f_min, f_max, direction) where T <: AbstractLPForm
 end
 
 "Constraint: compressor ratios on a new compressor--not applicable for MIP models-not applicable for LP models"
@@ -87,6 +105,10 @@ end
 
 "Constraint: Constraints which define pressure drop across a pipe "
 function constraint_pipe_pressure(gm::GenericGasModel{T}, n::Int, k, i, j, pd_min, pd_max) where T <: AbstractLPForm
+end
+
+"Constraint: Constraints which define pressure drop across a resistor "
+function constraint_resistor_pressure(gm::GenericGasModel{T}, n::Int, k, i, j, pd_min, pd_max) where T <: AbstractLPForm
 end
 
 "Constraint: constraints on pressure drop across an expansion pipe"