+++ src/form/mi/misocp.jl
@@ -55,7 +55,7 @@ function variable_mass_flow_ne(gm::GenericGasModel{T}, n::Int=gm.cnw; bounded::B
 end
 
 " Weymouth equation for an undirected pipe "
-function constraint_weymouth(gm::GenericGasModel{T}, n::Int, k, i, j, f_min, f_max, w, pd_min, pd_max) where T <: AbstractMISOCPForm
+function constraint_pipe_weymouth(gm::GenericGasModel{T}, n::Int, k, i, j, f_min, f_max, w, pd_min, pd_max) where T <: AbstractMISOCPForm
     y = var(gm,n,:y,k)
     pi = var(gm,n,:p,i)
     pj = var(gm,n,:p,j)
@@ -67,10 +67,31 @@ function constraint_weymouth(gm::GenericGasModel{T}, n::Int, k, i, j, f_min, f_m
    add_constraint(gm, n, :weymouth3, k, @constraint(gm.model, l <= pj - pi + pd_max*(2*y)))
    add_constraint(gm, n, :weymouth4, k, @constraint(gm.model, l <= pi - pj + pd_min*(2*y-2)))
    add_constraint(gm, n, :weymouth5, k, @constraint(gm.model, w*l >= f^2))
+
+   add_constraint(gm, n, :weymouth6, k, @constraint(gm.model, w*l <= f_max * f + (1-y) * (abs(f_min*f_max) + f_min^2)))
+   add_constraint(gm, n, :weymouth7, k, @constraint(gm.model, w*l <= f_min * f + y     * (abs(f_min*f_max) + f_max^2)))
+end
+
+" Weymouth equation for an undirected pipe "
+function constraint_resistor_weymouth(gm::GenericGasModel{T}, n::Int, k, i, j, f_min, f_max, w, pd_min, pd_max) where T <: AbstractMISOCPForm
+    y = var(gm,n,:y,k)
+    pi = var(gm,n,:p,i)
+    pj = var(gm,n,:p,j)
+    l  = var(gm,n,:l,k)
+    f  = var(gm,n,:f,k)
+
+   add_constraint(gm, n, :weymouth1, k, @constraint(gm.model, l >= pj - pi + pd_min*(2*y)))
+   add_constraint(gm, n, :weymouth2, k, @constraint(gm.model, l >= pi - pj + pd_max*(2*y-2)))
+   add_constraint(gm, n, :weymouth3, k, @constraint(gm.model, l <= pj - pi + pd_max*(2*y)))
+   add_constraint(gm, n, :weymouth4, k, @constraint(gm.model, l <= pi - pj + pd_min*(2*y-2)))
+   add_constraint(gm, n, :weymouth5, k, @constraint(gm.model, w*l >= f^2))
+
+   add_constraint(gm, n, :weymouth6, k, @constraint(gm.model, w*l <= f_max * f + (1-y) * (abs(f_min*f_max) + f_min^2)))
+   add_constraint(gm, n, :weymouth7, k, @constraint(gm.model, w*l <= f_min * f + y     * (abs(f_min*f_max) + f_max^2)))
 end
 
 "Weymouth equation with a pipe with one way flow"
-function constraint_weymouth_directed(gm::GenericGasModel{T}, n::Int, k, i, j, w, direction) where T <: AbstractMISOCPForm
+function constraint_pipe_weymouth_directed(gm::GenericGasModel{T}, n::Int, k, i, j, w, f_min, f_max, direction) where T <: AbstractMISOCPForm
     pi = var(gm,n,:p,i)
     pj = var(gm,n,:p,j)
     l  = var(gm,n,:l,k)
@@ -78,10 +99,31 @@ function constraint_weymouth_directed(gm::GenericGasModel{T}, n::Int, k, i, j, w
 
     add_constraint(gm, n, :weymouth1, k, @constraint(gm.model, l == direction * (pi - pj)))
     add_constraint(gm, n, :weymouth5, k, @constraint(gm.model, w*l >= f^2))
+    if (direction == 1)
+        add_constraint(gm, n, :weymouth6, k, @constraint(gm.model, w*l <= f_max * f))
+    else
+        add_constraint(gm, n, :weymouth7, k, @constraint(gm.model, w*l <= f_min * f))
+    end
+end
+
+"Weymouth equation with a resistor with one way flow"
+function constraint_resistor_weymouth_directed(gm::GenericGasModel{T}, n::Int, k, i, j, w, f_min, f_max, direction) where T <: AbstractMISOCPForm
+    pi = var(gm,n,:p,i)
+    pj = var(gm,n,:p,j)
+    l  = var(gm,n,:l,k)
+    f  = var(gm,n,:f,k)
+
+    add_constraint(gm, n, :weymouth1, k, @constraint(gm.model, l == direction * (pi - pj)))
+    add_constraint(gm, n, :weymouth5, k, @constraint(gm.model, w*l >= f^2))
+    if (direction == 1)
+        add_constraint(gm, n, :weymouth6, k, @constraint(gm.model, w*l <= f_max * f))
+    else
+        add_constraint(gm, n, :weymouth7, k, @constraint(gm.model, w*l <= f_min * f))
+    end
 end
 
 "Weymouth equation for an undirected expansion pipe"
-function constraint_weymouth_ne(gm::GenericGasModel{T},  n::Int, k, i, j, w, f_min, f_max, pd_min, pd_max) where T <: AbstractMISOCPForm
+function constraint_pipe_weymouth_ne(gm::GenericGasModel{T},  n::Int, k, i, j, w, f_min, f_max, pd_min, pd_max) where T <: AbstractMISOCPForm
     y = var(gm,n,:y_ne,k)
     pi = var(gm,n,:p,i)
     pj = var(gm,n,:p,j)
@@ -94,10 +136,13 @@ function constraint_weymouth_ne(gm::GenericGasModel{T},  n::Int, k, i, j, w, f_m
     add_constraint(gm, n, :weymouth_ne3, k,  @constraint(gm.model, l <= pj - pi + pd_max*(2*y)))
     add_constraint(gm, n, :weymouth_ne4, k,  @constraint(gm.model, l <= pi - pj + pd_min*(2*y-2)))
     add_constraint(gm, n, :weymouth_ne5, k,  @constraint(gm.model, zp*w*l >= f^2))
+
+    add_constraint(gm, n, :weymouth_ne6, k, @constraint(gm.model, w*l <= f_max * f + (1-y) * (abs(f_min*f_max) + f_min^2) + (1-zp) * (abs(f_min*f_max) + f_min^2)))
+    add_constraint(gm, n, :weymouth_ne7, k, @constraint(gm.model, w*l <= f_min * f + y     * (abs(f_min*f_max) + f_max^2) + (1-zp) * (abs(f_min*f_max) + f_max^2)))
 end
 
 "Weymouth equation for expansion pipes with undirected expansion pipes"
-function constraint_weymouth_ne_directed(gm::GenericGasModel{T},  n::Int, k, i, j, w, pd_min, pd_max, direction) where T <:  AbstractMISOCPForm
+function constraint_pipe_weymouth_ne_directed(gm::GenericGasModel{T},  n::Int, k, i, j, w, pd_min, pd_max, f_min, f_max, direction) where T <:  AbstractMISOCPForm
     pi = var(gm,n,:p,i)
     pj = var(gm,n,:p,j)
     zp = var(gm,n,:zp,k)
@@ -106,4 +151,10 @@ function constraint_weymouth_ne_directed(gm::GenericGasModel{T},  n::Int, k, i, 
 
     add_constraint(gm, n, :weymouth_ne1, k, @constraint(gm.model, l == direction * (pi - pj)))
     add_constraint(gm, n, :weymouth_ne5, k, @constraint(gm.model, zp*w*l >= f^2))
+    if (direction == 1)
+        add_constraint(gm, n, :weymouth_ne6, k, @constraint(gm.model, w*l <= f_max * f + (1-zp) * (abs(f_min*f_max) + f_min^2)))
+    else
+        add_constraint(gm, n, :weymouth_ne7, k, @constraint(gm.model, w*l <= f_min * f + (1-zp) * (abs(f_min*f_max) + f_max^2)))
+    end
+
 end