+++ src/form/mip.jl
@@ -14,26 +14,36 @@ const MIPGasModel = GenericGasModel{StandardMIPForm} # the standard MIP model
 "default MIP constructor"
 MIPGasModel(data::Dict{String,Any}; kwargs...) = GenericGasModel(data, StandardMIPForm)
 
-#################################################################################################
-### Variables
-#################################################################################################
-
 ######################################################################################################
 ## Constraints
 ######################################################################################################
 
 "Constraint: Weymouth equation--not applicable for MIP models"
-function constraint_weymouth(gm::GenericGasModel{T}, n::Int, k, i, j, f_min, f_max, w, pd_min, pd_max) where T <: AbstractMIPForm
+function constraint_pipe_weymouth(gm::GenericGasModel{T}, n::Int, k, i, j, f_min, f_max, w, pd_min, pd_max) where T <: AbstractMIPForm
+    #TODO we could think about putting a polyhendra around the weymouth
+end
+
+"Constraint: Weymouth equation--not applicable for MIP models"
+function constraint_resistor_weymouth(gm::GenericGasModel{T}, n::Int, k, i, j, f_min, f_max, w, pd_min, pd_max) where T <: AbstractMIPForm
     #TODO we could think about putting a polyhendra around the weymouth
 end
 
 "Constraint: Weymouth equation with one way direction--not applicable for MIP models"
-function constraint_weymouth_directed(gm::GenericGasModel{T}, n::Int, k, i, j, w, direction) where T <: AbstractMIPForm
+function constraint_pipe_weymouth_directed(gm::GenericGasModel{T}, n::Int, k, i, j, w, f_min, f_max, direction) where T <: AbstractMIPForm
+    #TODO we could think about putting a polyhendra around the weymouth
+end
+
+"Constraint: Weymouth equation with one way direction--not applicable for MIP models"
+function constraint_resistor_weymouth_directed(gm::GenericGasModel{T}, n::Int, k, i, j, w, f_min, f_max, direction) where T <: AbstractMIPForm
     #TODO we could think about putting a polyhendra around the weymouth
 end
 
 " Constraint: constraints on pressure drop across where direction is constrained"
 function constraint_pipe_pressure_directed(gm::GenericGasModel{T}, n::Int, k, i, j, pd_min, pd_max) where T <: AbstractMIPForm
+end
+
+" Constraint: constraints on pressure drop across where direction is constrained"
+function constraint_resistor_pressure_directed(gm::GenericGasModel{T}, n::Int, k, i, j, pd_min, pd_max) where T <: AbstractMIPForm
 end
 
 " Constraint: Constraint on pressure drop across a short pipe--not applicable for MIP models"
@@ -61,15 +71,15 @@ function constraint_on_off_control_valve_pressure_directed(gm::GenericGasModel{T
 end
 
 "Constraint: Weymouth equation--not applicable for MIP models--not applicable for MIP models"
-function constraint_weymouth_ne(gm::GenericGasModel{T},  n::Int, k, i, j, w, f_min, f_max, pd_min, pd_max) where T <: AbstractMIPForm
+function constraint_pipe_weymouth_ne(gm::GenericGasModel{T},  n::Int, k, i, j, w, f_min, f_max, pd_min, pd_max) where T <: AbstractMIPForm
 end
 
 " Constraint: Pressure drop across an expansion pipe when direction is constrained--not applicable for MIP models"
-function constraint_pressure_ne_directed(gm::GenericGasModel{T}, n::Int, k, i, j, yp, yn) where T <: AbstractMIPForm
+function constraint_pipe_pressure_ne_directed(gm::GenericGasModel{T}, n::Int, k, i, j, yp, yn) where T <: AbstractMIPForm
 end
 
 "Constraint: Weymouth equation--not applicable for MIP models--not applicable for MIP models"
-function constraint_weymouth_ne_directed(gm::GenericGasModel{T},  n::Int, k, i, j, w, pd_min, pd_max, direction) where T <: AbstractMIPForm
+function constraint_pipe_weymouth_ne_directed(gm::GenericGasModel{T},  n::Int, k, i, j, w, pd_min, pd_max, f_min, f_max, direction) where T <: AbstractMIPForm
 end
 
 "Constraint: compressor ratios on a new compressor--not applicable for MIP models-not applicable for MIP models"
@@ -82,6 +92,10 @@ end
 
 "Constraint: Constraints which define pressure drop across a pipe "
 function constraint_pipe_pressure(gm::GenericGasModel{T}, n::Int, k, i, j, pd_min, pd_max) where T <: AbstractMIPForm
+end
+
+"Constraint: Constraints which define pressure drop across a resistor "
+function constraint_resistor_pressure(gm::GenericGasModel{T}, n::Int, k, i, j, pd_min, pd_max) where T <: AbstractMIPForm
 end
 
 "Constraint: constraints on pressure drop across an expansion pipe"