+++ src/form/nlp.jl
@@ -19,7 +19,16 @@ NLPGasModel(data::Dict{String,Any}; kwargs...) = GenericGasModel(data, StandardN
 ############################################################################################################
 
 "Weymouth equation with absolute value "
-function constraint_weymouth(gm::GenericGasModel{T}, n::Int, k, i, j, f_min, f_max, w, pd_min, pd_max) where T <: AbstractNLPForm
+function constraint_pipe_weymouth(gm::GenericGasModel{T}, n::Int, k, i, j, f_min, f_max, w, pd_min, pd_max) where T <: AbstractNLPForm
+    pi = var(gm,n,:p,i)
+    pj = var(gm,n,:p,j)
+    f  = var(gm,n,:f,k)
+    add_constraint(gm, n, :weymouth1, k, @NLconstraint(gm.model, w*(pi - pj) <= f * abs(f)))
+    add_constraint(gm, n, :weymouth2, k, @NLconstraint(gm.model, w*(pi - pj) >= f * abs(f)))
+end
+
+"Weymouth equation with absolute value "
+function constraint_resistor_weymouth(gm::GenericGasModel{T}, n::Int, k, i, j, f_min, f_max, w, pd_min, pd_max) where T <: AbstractNLPForm
     pi = var(gm,n,:p,i)
     pj = var(gm,n,:p,j)
     f  = var(gm,n,:f,k)
@@ -28,7 +37,22 @@ function constraint_weymouth(gm::GenericGasModel{T}, n::Int, k, i, j, f_min, f_m
 end
 
 "Weymouth equation with one way direction"
-function constraint_weymouth_directed(gm::GenericGasModel{T}, n::Int, k, i, j, w, directed) where T <: AbstractNLPForm
+function constraint_pipe_weymouth_directed(gm::GenericGasModel{T}, n::Int, k, i, j, w, f_min, f_max, directed) where T <: AbstractNLPForm
+    pi = var(gm,n,:p,i)
+    pj = var(gm,n,:p,j)
+    f  = var(gm,n,:f,k)
+
+    if directed == 1
+        add_constraint(gm, n, :weymouth_ne1, k, @constraint(gm.model, w*(pi - pj) >= f^2))
+        add_constraint(gm, n, :weymouth_ne2, k, @constraint(gm.model, w*(pi - pj) <= f^2))
+    else
+        add_constraint(gm, n, :weymouth_ne3, k, @constraint(gm.model, w*(pj - pi) >= f^2))
+        add_constraint(gm, n, :weymouth_ne4, k, @constraint(gm.model, w*(pj - pi) <= f^2))
+    end
+end
+
+"Weymouth equation with one way direction"
+function constraint_resistor_weymouth_directed(gm::GenericGasModel{T}, n::Int, k, i, j, w, f_min, f_max, directed) where T <: AbstractNLPForm
     pi = var(gm,n,:p,i)
     pj = var(gm,n,:p,j)
     f  = var(gm,n,:f,k)
@@ -47,7 +71,7 @@ end
 ############################################################################################################
 
 "Weymouth equation for directed expansion pipes"
-function constraint_weymouth_ne_directed(gm::GenericGasModel{T},  n::Int, k, i, j, w, pd_min, pd_max, direction) where T <: AbstractNLPForm
+function constraint_pipe_weymouth_ne_directed(gm::GenericGasModel{T},  n::Int, k, i, j, w, pd_min, pd_max, f_min, f_max, direction) where T <: AbstractNLPForm
     pi = var(gm,n,:p,i)
     pj = var(gm,n,:p,j)
     zp = var(gm,n,:zp,k)
@@ -64,7 +88,7 @@ function constraint_weymouth_ne_directed(gm::GenericGasModel{T},  n::Int, k, i, 
 end
 
 " Weymouth equation for an undirected expansion pipe "
-function constraint_weymouth_ne(gm::GenericGasModel{T},  n::Int, k, i, j, w, f_min, f_max, pd_min, pd_max) where T <: AbstractNLPForm
+function constraint_pipe_weymouth_ne(gm::GenericGasModel{T},  n::Int, k, i, j, w, f_min, f_max, pd_min, pd_max) where T <: AbstractNLPForm
     pi = var(gm,n,:p,i)
     pj = var(gm,n,:p,j)
     zp = var(gm,n,:zp,k)
@@ -117,5 +141,5 @@ function constraint_on_off_control_valve_pressure(gm::GenericGasModel{T}, n::Int
     #TODO this constraint is only valid if max_ratio = 1
     add_constraint(gm, n, :control_valve_pressure_drop1, k, @constraint(gm.model, pj - max_ratio^2*pi <= (1-v)*j_pmax^2))
     add_constraint(gm, n, :control_valve_pressure_drop2, k, @constraint(gm.model, min_ratio^2*pi - pj <= (1-v)*(min_ratio*i_pmax^2)))
-    add_constraint(gm, n, :control_valve_pressure_drop3, k, @NLconstraint(gm.model, f * (1-pj/pi) >= (1-zc) * f_max * (1-j_pmax^2/i_pmin^2)))
+    add_constraint(gm, n, :control_valve_pressure_drop3, k, @NLconstraint(gm.model, f * (1-pj/pi) >= (1-v) * f_max * (1-j_pmax^2/i_pmin^2)))
 end