+++ src/prob/nels.jl
@@ -37,20 +37,20 @@ function post_nels(gm::GenericGasModel)
     for i in ids(gm,:pipe)
         constraint_pipe_pressure(gm, i)
         constraint_pipe_mass_flow(gm,i)
-        constraint_weymouth(gm,i)
+        constraint_pipe_weymouth(gm,i)
     end
 
     for i in ids(gm,:resistor)
-        constraint_pipe_pressure(gm, i)
-        constraint_pipe_mass_flow(gm,i)
-        constraint_weymouth(gm,i)
+        constraint_resistor_pressure(gm, i)
+        constraint_resistor_mass_flow(gm,i)
+        constraint_resistor_weymouth(gm,i)
     end
 
     for i in ids(gm,:ne_pipe)
         constraint_pipe_pressure_ne(gm, i)
         constraint_pipe_ne(gm, i)
         constraint_pipe_mass_flow_ne(gm,i)
-        constraint_weymouth_ne(gm, i)
+        constraint_pipe_weymouth_ne(gm, i)
     end
 
     for i in ids(gm, :short_pipe)
@@ -119,39 +119,39 @@ function post_nels_directed(gm::GenericGasModel)
     for i in ids(gm,:undirected_pipe)
         constraint_pipe_pressure(gm, i)
         constraint_pipe_mass_flow(gm,i)
-        constraint_weymouth(gm,i)
+        constraint_pipe_weymouth(gm,i)
     end
 
     for i in ids(gm,:undirected_resistor)
-        constraint_pipe_pressure(gm, i)
-        constraint_pipe_mass_flow(gm,i)
-        constraint_weymouth(gm,i)
+        constraint_resistor_pressure(gm, i)
+        constraint_resistor_mass_flow(gm,i)
+        constraint_resistor_weymouth(gm,i)
     end
 
     for i in ids(gm,:directed_pipe)
         constraint_pipe_pressure_directed(gm, i)
         constraint_pipe_mass_flow_directed(gm, i)
-        constraint_weymouth_directed(gm, i)
+        constraint_pipe_weymouth_directed(gm, i)
     end
 
     for i in ids(gm,:directed_resistor)
-        constraint_pipe_pressure_directed(gm, i)
-        constraint_pipe_mass_flow_directed(gm, i)
-        constraint_weymouth_directed(gm, i)
+        constraint_resistor_pressure_directed(gm, i)
+        constraint_resistor_mass_flow_directed(gm, i)
+        constraint_resistor_weymouth_directed(gm, i)
     end
 
     for i in ids(gm,:undirected_ne_pipe)
         constraint_pipe_pressure_ne(gm, i)
         constraint_pipe_ne(gm, i)
         constraint_pipe_mass_flow_ne(gm,i)
-        constraint_weymouth_ne(gm, i)
+        constraint_pipe_weymouth_ne(gm, i)
     end
 
     for i in ids(gm,:directed_ne_pipe)
-        constraint_pressure_ne_directed(gm, i)
+        constraint_pipe_pressure_ne_directed(gm, i)
         constraint_pipe_mass_flow_ne_directed(gm, i)
         constraint_pipe_ne(gm, i)
-        constraint_weymouth_ne_directed(gm, i)
+        constraint_pipe_weymouth_ne_directed(gm, i)
     end
 
     for i in ids(gm, :undirected_short_pipe)