+++ CHANGELOG.md
@@ -7,6 +7,9 @@ GasModels.jl Change Log
 
 ### v0.4.3
 - Renamed various function names for consistency (breaking)
+- Replaced yn and yp variables with a single variable y
+- Broke out resistor constraints from pipe constraints (breaking)
+- Added upper bounds on second order code relaxations
 
 ### v0.4.2
 - Implement nlp version(s) of the problems